Fixed a bug related to settings configuration
Settings enumerations used the hex prefix `0x` for register 
values. That was fine for most cases, as it were `0` and `1` values. 
However, in cases where the value should have been 3 in binary 
it was 0x11 instead of the binary prefix 0b11. This caused the output 
data rate to never be properly configured for the frequencies 20Hz 
and 100hz.

To avoid this issue in the future, all config values prefix have been 
replaced with the binary 0b0 prefix.