Add support for Offset Drift Function
Add support for the "Offset Drift Function" of the HSCDTD008A
sensor. This allows the user to apply a drift correction for
each axis. This change is added to both the C Driver and
the Arduino Wrapper layer. This feature utilizes the hardware 
capabilities of the sensor, the offset is not applied in software.