Fixed no data issue
Resolved an issue that caused data to not be available of TCS or OCL 
was used. The issue was caused by incorrectly assigning the mode value
to the state value of the virtual device. An additional safeguard has
been added that explicitly configures the device in the force state during
initialization.